# docs/swagger.yaml
definitions:
...
  controller.ResponseCodeType:
    enum:
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 1009
    - 1010
    - 1011
    - 1012
    - 1013
    - 1014
    - 1015
    - 1016
    - 1017
    - 1018
    type: integer
    x-enum-varnames:
    - CodeSuccess
    - CodeParameterInvalid
    - CodeUserNotExist
    - CodeUserExist
    - CodePasswordError
    - CodeServerBusy
    - CodeTokenInvaild
    - CodeTokenEmpty
    - CodeNeedReLogin
    - CodeCommunityIdInvalid
    - CodeVerifyCodeInvaild
    - CodeArticleTitleExisted
    - CodeArticleNotExisted
    - CodeHaveArticleInClass
    - CodeTagExisted
    - CodeTagNotExisted
    - CodeClassNotExisted
    - CodeClassNameExisted
    - CodeCreateCommentFailed
  controller.ResponseData:
    properties:
      code:
        $ref: '#/definitions/controller.ResponseCodeType'
      data:
        description: omitempty 该字段为空时忽略
      msg: {}
    type: object
...
  models.NewArticleFormsParams:
    properties:
      class_id:
        description: 文章所属的分类 - 比如教程，分享等
        example: "0"
        type: string
      content:
        description: 文章内容
        type: string
      en_comment:
        description: 是否允许评论 false-不允许 true-允许
        type: boolean
      image:
        description: 文章缩略图
        type: string
      privilege:
        allOf:
        - $ref: '#/definitions/models.PrivilegeType'
        description: 文章权限 0-公开 1-私有
      status:
        description: 文章状态 0-草稿 1-发布 2-删除
        type: integer
      summary:
        description: 文章的摘要信息
        type: string
      tag_id_list:
        description: 文章标签-标签可以为空
        items:
          $ref: '#/definitions/models.TagFormsParams'
        type: array
      title:
        description: 文章标题
        type: string
      top_flag:
        description: 是否置顶标志 false-没有置顶 true-置顶
        type: boolean
    required:
    - class_id
    - content
    - title
    type: object
  models.PrivilegeType:
    enum:
    - 1
    - 2
    type: integer
    x-enum-comments:
      PrivilegePrivte: 2- 当前文章私有，客户端不可见
      PrivilegePublic: 1 - 当前文章是公开的，客户端可见（默认）
    x-enum-varnames:
    - PrivilegePublic
    - PrivilegePrivte
...
  models.SoftDeleteArticleParams:
    properties:
      del_flag:
        description: true - 删除  false - 恢复
        type: boolean
      ids:
        description: 待删除的文章ID列表
        items:
          type: integer
        type: array
    required:
    - del_flag
    - ids
    type: object
...
  models.TagFormsParams:
    properties:
      id:
        description: tag id
        example: "0"
        type: string
    required:
    - id
    type: object
  models.UpdateArticleFormsParams:
    properties:
      article_id:
        example: "0"
        type: string
      class_id:
        description: 文章所属的分类 - 比如教程，分享等
        example: "0"
        type: string
      content:
        description: 文章内容
        type: string
      en_comment:
        description: 是否允许评论 false-不允许 true-允许
        type: boolean
      image:
        description: 文章缩略图
        type: string
      privilege:
        allOf:
        - $ref: '#/definitions/models.PrivilegeType'
        description: 文章权限 0-公开 1-私有
      status:
        description: 文章状态 0-草稿 1-发布 2-删除
        type: integer
      summary:
        description: 文章的摘要信息
        type: string
      tag_id_list:
        description: 文章标签-标签可以为空
        items:
          $ref: '#/definitions/models.TagFormsParams'
        type: array
      title:
        description: 文章标题
        type: string
      top_flag:
        description: 是否置顶标志 false-没有置顶 true-置顶
        type: boolean
    required:
    - article_id
    - class_id
    - content
    - title
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: NgBlog Go博客项目 API 接口文档
  title: NgBlog
  version: "1.0"
paths:
  /article:
    post:
      consumes:
      - application/json
      description: 通过该接口可以创建文章
      parameters:
      - description: Bearer token(jwt)
        in: header
        name: Authorization
        required: true
        type: string
      - description: 创建文章的参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.NewArticleFormsParams'
      produces:
      - application/json
      responses:
        "200":
          description: code=200表示成功其余失败
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 创建文章的接口
      tags:
      - 文章相关接口
    put:
      consumes:
      - application/json
      description: 通过该接口可以更新指定的文章
      parameters:
      - description: Bearer token(jwt)
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新文章的参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.UpdateArticleFormsParams'
      produces:
      - application/json
      responses:
        "200":
          description: code=200表示成功其余失败
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 更新文章的接口
      tags:
      - 文章相关接口
  /article/:id:
    delete:
      consumes:
      - application/json
      description: 通过该接口可以删除指定的文章
      parameters:
      - description: Bearer token(jwt)
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code=200表示成功其余失败
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 删除单篇文章的接口
      tags:
      - 文章相关接口
    get:
      consumes:
      - application/json
      description: 通过该接口可以查询文章详细信息
      parameters:
      - description: Bearer token(jwt)
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code=200表示成功其余失败
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 通过文章ID查询文章详细信息的接口
      tags:
      - 文章相关接口
  /articles:
    delete:
      consumes:
      - application/json
      description: 通过该接口可以删除指定的文章 - 软删除
      parameters:
      - description: Bearer token(jwt)
        in: header
        name: Authorization
        required: true
        type: string
      - description: 删除文章的参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.SoftDeleteArticleParams'
      produces:
      - application/json
      responses:
        "200":
          description: code=200表示成功,其余失败
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 删除时,只会更新其状态为删除状态,不会删除数据库中的数据，恢复时将其恢复为指定状态
      tags:
      - 文章相关接口
    get:
      consumes:
      - application/json
      description: 通过该接口可以获得当前的所有文章
      parameters:
      - description: Bearer token(jwt)
        in: header
        name: Authorization
        required: true
        type: string
      - description: page
        in: query
        name: p
        type: integer
      - description: size
        in: query
        name: s
        type: integer
      - description: class
        in: query
        name: c
        type: string
      - description: tag
        in: query
        name: t
        type: string
      - description: name
        in: query
        name: "n"
        type: string
      - description: status
        in: query
        name: st
        type: string
      - description: privilege
        in: query
        name: pr
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 获取所有文章的接口
      tags:
      - 文章相关接口
...
swagger: "2.0"